Implement displaying huge article body
If i had rendered the article body as a whole, it takes a while to load the body

To fix this, we splited the huge text body into multiple smaller chunks, place a Flatlist and pass the array of paragraphs to it.

FYI:
Fir our article that has web links in them, i made the text highlightable so that the user can copy the url. Another option will be to check if the text includes a url then you will render a different component other than Text
- https://www.codegrepper.com/code-examples/javascript/how+to+check+string+contains+a+string+in+react+native
- https://stackoverflow.com/questions/34933860/react-native-how-to-detect-link-form-string-and-convert-to-a-link
- NOTE: the selectable is not working as much as i want for many reasons, so this could be the best option

Other userFul resources on optimizing flatlist
- https://forums.expo.io/t/how-to-load-more-than-1000-items-in-a-flat-list/29112
- https://aboutreact.com/infinite-list-view/
- https://reactnative.dev/docs/optimizing-flatlist-configuration